fixes #4 #47 #45 #40 #37

+++ bash-scripts/hooks_generator.sh
@@ -9,4 +9,10 @@ sed -i 's/_INSERT_VM_NAME_/$1/' tmp/hooks/release/end/revert.sh
 
 
 mkdir -p /etc/libvirt/hooks/qemu.d/$1
-cp -R tmp/hooks /etc/libvirt/hooks/qemu.d/$1/hooks
+
+ls /etc/libvirt/hooks/qemu.d/$1/hooks
+hooks_exists=$?
+if [ hooks_exists != 0 ]
+then
+	cp -R tmp/hooks /etc/libvirt/hooks/qemu.d/$1/hooks
+fi
